+++ Server/Diagnostics/TimerProfile.cs
@@ -5,7 +5,7 @@
  *   copyright            : (C) The RunUO Software Team
  *   email                : info@runuo.com
  *
- *   $Id: PacketProfile.cs 4 2006-06-15 04:28:39Z mark $
+ *   $Id$
  *
  ***************************************************************************/
 
@@ -47,9 +47,9 @@ namespace Server.Diagnostics {
 			return prof;
 		}
 
-		private int _created, _started, _stopped;
+		private long _created, _started, _stopped;
 
-		public int Created {
+		public long Created {
 			get {
 				return _created;
 			}
@@ -58,7 +58,7 @@ namespace Server.Diagnostics {
 			}
 		}
 
-		public int Started {
+		public long Started {
 			get {
 				return _started;
 			}
@@ -67,7 +67,7 @@ namespace Server.Diagnostics {
 			}
 		}
 
-		public int Stopped {
+		public long Stopped {
 			get {
 				return _stopped;
 			}
